Support `tw-curator.disabled: true` case
Users of this library does not have to introduce Zookeeper as a dependency. E.g. `Wise Kafka Processor` can support database cleanup that depends on leader selection provided by thislibrary. But `Wise Kafka Processor` can also be used without a DB dependency, in which case we want to avoid introducing the dependency on Zookeeper.